We can now change all scent names to Swedish
Essentially, what you have to be careful of now is that the name on the tag, matches the name of the sprite (preferably let's continue to write those in english). Unity will now check for a match between the name of the sprite and the name of the tag to create a new scent card. THEREFORE, we can use whatever accents or spaces we want on the Name of our scent cards now :)

Co-Authored-By: Agustin Guerrero <16386119+Dibujaman@users.noreply.github.com>